[Android] Fix onReceivedHttpAuthRequest() was not called issue.
This patch is to fix the issue about onReceivedHttpAuthRequest() was not
called when this function was overridden in class of XWalkResourceClient.
Modify the attribute of XWalkHttpAuthHandlerInternal to generate right
parameter for XWalkResourceClientBridge.
Move test case from XWalkViewInternal to XWalkView.

BUG=XWALK-5117

(cherry picked from commit b89f5dd217fddb9b7f8ba84ca89383a09c37f8eb)
