+++ test/android/core_internal/javatests/src/org/xwalk/core/internal/xwview/test/TestHelperBridge.java
@@ -127,20 +127,6 @@ class TestHelperBridge {
         }
     }
 
-    public class OnReceivedHttpAuthRequestHelper extends CallbackHelper {
-        private String mHost;
-
-        public String getHost() {
-            assert getCallCount() > 0;
-            return mHost;
-        }
-
-        public void notifyCalled(String host) {
-            mHost = host;
-            notifyCalled();
-        }
-    }
-
     private String mChangedTitle;
     private final OnPageStartedHelper mOnPageStartedHelper;
     private final OnPageFinishedHelper mOnPageFinishedHelper;
@@ -151,7 +137,6 @@ class TestHelperBridge {
     private final OnTitleUpdatedHelper mOnTitleUpdatedHelper;
     private final ShouldInterceptLoadRequestHelper mShouldInterceptLoadRequestHelper;
     private final OnLoadStartedHelper mOnLoadStartedHelper;
-    private final OnReceivedHttpAuthRequestHelper mOnReceivedHttpAuthRequestHelper;
 
     public TestHelperBridge() {
         mOnPageStartedHelper = new OnPageStartedHelper();
@@ -161,7 +146,6 @@ class TestHelperBridge {
         mOnTitleUpdatedHelper = new OnTitleUpdatedHelper();
         mShouldInterceptLoadRequestHelper = new ShouldInterceptLoadRequestHelper();
         mOnLoadStartedHelper = new OnLoadStartedHelper();
-        mOnReceivedHttpAuthRequestHelper = new OnReceivedHttpAuthRequestHelper();
     }
 
     public OnPageStartedHelper getOnPageStartedHelper() {
@@ -190,10 +174,6 @@ class TestHelperBridge {
 
     public OnLoadStartedHelper getOnLoadStartedHelper() {
         return mOnLoadStartedHelper;
-    }
-
-    public OnReceivedHttpAuthRequestHelper getOnReceivedHttpAuthRequestHelper() {
-        return mOnReceivedHttpAuthRequestHelper;
     }
 
     public void onTitleChanged(String title) {
@@ -225,9 +205,5 @@ class TestHelperBridge {
 
     public void onLoadStarted(String url) {
         mOnLoadStartedHelper.notifyCalled(url);
-    }
-
-    public void onReceivedHttpAuthRequest(String host) {
-        mOnReceivedHttpAuthRequestHelper.notifyCalled(host);
     }
 }
