+++ tools/reflection_generator/reflection_generator.py
@@ -32,6 +32,7 @@ CLASSES_TO_BE_PROCESS = [
     'XWalkExtensionInternal',
     'XWalkGetBitmapCallbackInternal',
     'XWalkHttpAuthHandlerInternal',
+    'XWalkHttpAuthInternal',
     'XWalkViewInternal',
     'XWalkUIClientInternal',
     'XWalkResourceClientInternal',
